chat_application/requirements.txt
gitpython
pinecone-client
streamlit-chat
streamlit-extras
streamlit
langchain